+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { SettingsForm } from './SettingsForm';
+export { InfoPanel } from './InfoPanel';
+export { ResultsPanel } from './ResultsPanel';
+export { default as Timer } from './Timer';
